requirements.txt
amqp==5.2.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
asgiref==3.7.2 ; python_full_version >= "3.12.1" and python_version < "4.0"
attrs==23.2.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
billiard==4.2.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
black==23.12.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
celery==5.3.6 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
celery[redis]==5.3.6 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
certifi==2024.2.2 ; python_full_version >= "3.12.1" and python_version < "4.0"
cffi==1.16.0 ; python_full_version >= "3.12.1" and python_version < "4.0" and platform_python_implementation != "PyPy"
cfgv==3.4.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
charset-normalizer==3.3.2 ; python_full_version >= "3.12.1" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
click-plugins==1.1.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
click-repl==0.3.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
click==8.1.7 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0" and platform_system == "Windows"
cron-descriptor==1.4.3 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
cryptography==42.0.3 ; python_full_version >= "3.12.1" and python_version < "4.0"
defusedxml==0.7.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
distlib==0.3.8 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
django-celery-beat==2.6.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
django-cors-headers==4.3.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
django-filter==23.5 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
django-templated-mail==1.1.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
django-timezone-field==6.1.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
django==5.0.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
djangorestframework==3.14.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
djoser==2.2.2 ; python_full_version >= "3.12.1" and python_version < "4.0"
drf-spectacular==0.27.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
filelock==3.13.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
gunicorn==21.2.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
identify==2.5.34 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
idna==3.6 ; python_full_version >= "3.12.1" and python_version < "4.0"
inflection==0.5.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
isort==5.13.2 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
jsonschema-specifications==2023.12.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
jsonschema==4.21.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
kombu==5.3.5 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
mypy-extensions==1.0.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
nodeenv==1.8.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
oauthlib==3.2.2 ; python_full_version >= "3.12.1" and python_version < "4.0"
packaging==23.2 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
pathspec==0.12.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
pillow==10.2.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
platformdirs==4.2.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
pre-commit==3.6.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
prompt-toolkit==3.0.43 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
psycopg2-binary==2.9.9 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
pycparser==2.21 ; python_full_version >= "3.12.1" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyjwt==2.8.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
python-crontab==3.0.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
python-dateutil==2.9.0.post0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
python-dotenv==1.0.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
python3-openid==3.2.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
pytz==2024.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
pyyaml==6.0.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
redis==5.0.3 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
referencing==0.33.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
requests-oauthlib==1.3.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
requests==2.31.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
rpds-py==0.18.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
setuptools==69.1.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
six==1.16.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
social-auth-app-django==5.4.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
social-auth-core==4.5.3 ; python_full_version >= "3.12.1" and python_version < "4.0"
sqlparse==0.4.4 ; python_full_version >= "3.12.1" and python_version < "4.0"
tzdata==2024.1 ; python_full_version >= "3.12.1" and python_version < "4.0"
uritemplate==4.1.1 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
urllib3==2.2.0 ; python_full_version >= "3.12.1" and python_version < "4.0"
vine==5.1.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
virtualenv==20.25.0 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
wcwidth==0.2.13 ; python_full_version >= "3.12.1" and python_full_version < "4.0.0"
